Still stuck with floated block level elements
The event width at 100% totally screws up the elements that need to be shared, and I'm definitely not keen on doing some hackery with js to figure out where the overlap happens and change widths that way.

I'd actually prefer to leave the events not at 100% width than introduce js to solve this problem (which would more likely create more problems). There is a CSS solution for this, as a handful of StackOverflow articles would suggest. I just haven't had success with it. I think it's interesting that this is the first time in 11 years that I've crossed this problem. It's kind of fun.